refactor: option 패키지를 없애고 option과 관련된 파일들을 product 패키지에 통합
